Bug that might be linked to initialisation
